requirements.txt
streamlit==1.30.0
google-generativeai==0.8.5
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.21
jsonschema==4.19.0
